toolkit/content/tests/widgets/test_videocontrols_keyhandler.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Video controls test - KeyHandler</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="head.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>

<div id="content">
  <video id="video" controls preload="auto"></video>
</div>

<pre id="test">
<script class="testbody" type="application/javascript">
  SimpleTest.waitForExplicitFinish();
  const video = document.getElementById("video");

  const playButton = getElementWithinVideo(video, "playButton");
  const scrubber = getElementWithinVideo(video, "scrubber");
  const volumeStack = getElementWithinVideo(video, "volumeStack");

  // Setup video
  tests.push(done => {
    SpecialPowers.pushPrefEnv({"set": [["media.cache_size", 40000]]}, done);
  }, done => {
    video.src = "seek_with_sound.ogg";
    video.addEventListener("loadedmetadata", done);
  });

  // Bug 1350191, video should not seek while changing volume by
  // pressing up/down arrow key.
  tests.push(done => {
    video.addEventListener("play", done, { once: true });
    synthesizeMouseAtCenter(playButton, {});
  }, done => {
    video.addEventListener("seeked", done, { once: true });
    synthesizeMouseAtCenter(scrubber, {});
  }, done => {
    let counter = 0;
    let keys = ["KEY_ArrowDown", "KEY_ArrowDown", "KEY_ArrowUp", "KEY_ArrowDown", "KEY_ArrowUp", "KEY_ArrowUp"];

    video.addEventListener("seeked", () => ok(false, "should not trigger seeked event"));
    video.addEventListener("volumechange", () => {
      if (++counter === keys.length) {
        ok(true, "change volume by up/down arrow key without trigger 'seeked' event");
        done();
      }

      if (counter > keys.length) {
        ok(false, "trigger too much volumechange events");
      }
    });

    for (let key of keys) {
      synthesizeKey(key);
    }
  });

  tests.push(SimpleTest.finish);

  window.addEventListener("load", executeTests);
</script>
</pre>
</body>
</html>
